+++ CHANGES.rst
@@ -5,7 +5,8 @@
 4.3.0 (unreleased)
 ==================
 
-- Support options `use_datetime` and `use_builtin_types` of ``ServerProxy``.
+- Support options `use_datetime` (`Python >= 2.7`) and `use_builtin_types`
+  (`Python >= 3.5`) of ``ServerProxy``.
 
 
 4.2.0 (2019-12-05)